only port2 remove old entries

+++ src/app/components/new-entry-window/new-entry-window.component.ts
@@ -157,8 +157,9 @@ export class NewEntryWindowComponent {
     }
     this.newEntryButtonIn.nativeElement.disabled = true;
     this.newEntryButtonOut.nativeElement.disabled = true;
-
-    await EntryService.removeOldEntries();
+    if (this.porta1RadioButtonIsChecked === false) {
+      await EntryService.removeOldEntries();
+    }
   }
 
   ngAfterViewInit() {
